Update index.js
p tags are used to generate the description when other methods fail.  However, there are cases where the p tag may have internal  script tags and the javascript is interpreted as text  and so is used for  the  description; this is an obvious error. This fix removes any p tag  with internal script tags from consideration when generating a description. Follett internal reference D-21669.